+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [4.0.1](https://github.com/rizen/pixi-glyphs/compare/v5.0.0...v4.0.1) (2025-09-19)
 
+- Fixed: PixiJS Deprecation Warning: Texture.baseTexture is now Texture.source
+
 ## 4.0.0 (2025-09-19)
 
 - Forked from pixi-tagged-text and updated to use pixi.js v8.